Move /essentials to tree command
