+++ Essentials/src/main/java/com/earth2me/essentials/utils/VersionUtil.java
@@ -9,6 +9,8 @@ import net.ess3.nms.refl.ReflUtil;
 import org.bukkit.Bukkit;
 
 import java.lang.reflect.Method;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.regex.Matcher;
@@ -47,6 +49,26 @@ public final class VersionUtil {
     private static final Set<BukkitVersion> supportedVersions = ImmutableSet.of(v1_8_8_R01, v1_9_4_R01, v1_10_2_R01, v1_11_2_R01, v1_12_2_R01, v1_13_2_R01, v1_14_4_R01, v1_15_2_R01, v1_16_5_R01, v1_17_1_R01, v1_18_2_R01, v1_19_4_R01, v1_20_6_R01, v1_21_8_R01);
 
     public static final boolean PRE_FLATTENING = VersionUtil.getServerBukkitVersion().isLowerThan(VersionUtil.v1_13_0_R01);
+
+    public static final List<String> officialPlugins = Arrays.asList(
+            "EssentialsAntiBuild",
+            "EssentialsChat",
+            "EssentialsDiscord",
+            "EssentialsDiscordLink",
+            "EssentialsGeoIP",
+            "EssentialsProtect",
+            "EssentialsSpawn",
+            "EssentialsXMPP"
+    );
+    public static final List<String> warnPlugins = Arrays.asList(
+            "PermissionsEx",
+            "GroupManager",
+            "bPermissions",
+
+            // Brain-dead chat signing bypass that break EssentialsChat
+            "NoChatReports",
+            "NoEncryption"
+    );
 
     private static final Map<String, SupportStatus> unsupportedServerClasses;
